More robust path normalization
